requirements-conda.txt
fmpy-base
pytest
jupyter_core
requests